pyproject.toml
[tool.poetry]
name = "fmp-loader"
version = "0.1.0"
description = ""
authors = ["Weeruhputt Supsohmboon"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
SQLAlchemy = "^1.4.40"
dataclasses-json = "^0.5.7"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^8.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
